Move test fixtures used in MFS to MF proper
