+++ tests/dataflow_plan_to_svg.py
@@ -5,8 +5,8 @@ import os
 from _pytest.fixtures import FixtureRequest
 
 from metricflow.dag.dag_visualization import DagGraphT, render_via_graphviz
+from metricflow.test_helpers import snapshot_path_prefix
 from tests.fixtures.setup_fixtures import MetricFlowTestConfiguration
-from tests.snapshot_utils import snapshot_path_prefix
 
 
 def display_graph_if_requested(
