+++ tests/snapshot_utils.py
@@ -1,11 +1,8 @@
 from __future__ import annotations
 
-import difflib
 import logging
-import os
 import re
-import webbrowser
-from typing import Any, Callable, List, Optional, Tuple, TypeVar
+from typing import Any, Callable, Optional, Tuple, TypeVar
 
 import tabulate
 from _pytest.fixtures import FixtureRequest
@@ -18,6 +15,7 @@ from metricflow.model.semantics.linkable_spec_resolver import LinkableElementSet
 from metricflow.naming.object_builder_scheme import ObjectBuilderNamingScheme
 from metricflow.protocols.sql_client import SqlClient, SqlEngine
 from metricflow.specs.specs import InstanceSpecSet, LinkableSpecSet
+from metricflow.test_helpers import assert_snapshot_text_equal
 from tests.fixtures.setup_fixtures import MetricFlowTestConfiguration, check_sql_engine_snapshot_marker
 
 logger = logging.getLogger(__name__)
@@ -62,64 +60,7 @@ def replace_dataset_id_hash(text: str) -> str:
     return text
 
 
-def snapshot_path_prefix(
-    request: FixtureRequest,
-    snapshot_group: str,
-    snapshot_id: str,
-    additional_sub_directories: Tuple[str, ...] = (),
-) -> str:
-    """Returns a path prefix that can be used to build filenames for files associated with the snapshot.
-
-    The snapshot prefix is generated from the name of the test file, the name of the test, name of the snapshot class,
-    and the name of the snapshot.
-
-    e.g.
-    .../snapshots/test_file.py/DataflowPlan/test_name__plan1
-
-    which can be used to construct paths like
-
-    .../snapshots/test_file.py/DataflowPlan/test_name__plan1.xml
-    .../snapshots/test_file.py/DataflowPlan/test_name__plan1.svg
-    """
-    test_name = request.node.name
-
-    snapshot_file_name_parts = []
-    # Parameterized test names look like 'test_case[some_param]'. "[" and "]" are annoying to deal with in the shell,
-    # so replace them with dunders.
-    snapshot_file_name_parts.extend(re.split(r"[\[\]]", test_name))
-    # A trailing ] will produce an empty string in the list, so remove that.
-    snapshot_file_name_parts = [part for part in snapshot_file_name_parts if len(part) > 0]
-    snapshot_file_name_parts.append(snapshot_id)
-
-    snapshot_file_name = "__".join(snapshot_file_name_parts)
-
-    path_items: List[str] = []
-
-    test_file_path_items = os.path.normpath(request.node.fspath).split(os.sep)
-    test_file_name = test_file_path_items[-1]
-    # Default to where this is defined, but use more appropriate directories if found.
-    test_directory_root_index = -1
-    for i, path_item in enumerate(test_file_path_items):
-        if path_item in ("tests", "metricflow"):
-            test_directory_root_index = i + 1
-
-    path_to_store_snapshots = os.sep.join(test_file_path_items[:test_directory_root_index])
-    path_items.extend([path_to_store_snapshots, "snapshots", test_file_name, snapshot_group])
-
-    if additional_sub_directories:
-        path_items.extend(additional_sub_directories)
-    path_items.append(snapshot_file_name)
-
-    return os.path.abspath(os.path.join(*path_items))
-
-
 PlanT = TypeVar("PlanT", bound=MetricFlowDag)
-
-
-def _exclude_lines_matching_regex(file_contents: str, exclude_line_regex: str) -> str:
-    """Removes lines from file contents if the line matches exclude_regex."""
-    compiled_regex = re.compile(exclude_line_regex)
-    return "\n".join([line for line in file_contents.split("\n") if not compiled_regex.match(line)])
 
 
 def assert_plan_snapshot_text_equal(
@@ -154,77 +95,6 @@ def assert_plan_snapshot_text_equal(
         incomparable_strings_replacement_function=incomparable_strings_replacement_function,
         additional_sub_directories_for_snapshots=additional_sub_directories_for_snapshots,
     )
-
-
-def assert_snapshot_text_equal(
-    request: FixtureRequest,
-    mf_test_configuration: MetricFlowTestConfiguration,
-    group_id: str,
-    snapshot_id: str,
-    snapshot_text: str,
-    snapshot_file_extension: str,
-    exclude_line_regex: Optional[str] = None,
-    incomparable_strings_replacement_function: Optional[Callable[[str], str]] = None,
-    additional_sub_directories_for_snapshots: Tuple[str, ...] = (),
-) -> None:
-    """Similar to assert_plan_snapshot_text_equal(), but with more controls on how the snapshot paths are generated."""
-    file_path = (
-        snapshot_path_prefix(
-            request=request,
-            snapshot_group=group_id,
-            snapshot_id=snapshot_id,
-            additional_sub_directories=additional_sub_directories_for_snapshots,
-        )
-        + snapshot_file_extension
-    )
-
-    if incomparable_strings_replacement_function is not None:
-        snapshot_text = incomparable_strings_replacement_function(snapshot_text)
-
-    # Add a new line at the end of the file so that PRs don't show the "no newline" symbol on Github.
-    if len(snapshot_text) > 1 and snapshot_text[-1] != "\n":
-        snapshot_text = snapshot_text + "\n"
-
-    # If we are in overwrite mode, make a new plan:
-    if mf_test_configuration.overwrite_snapshots:
-        # Create parent directory for the plan text files.
-        os.makedirs(os.path.dirname(file_path), exist_ok=True)
-        with open(file_path, "w") as snapshot_text_file:
-            snapshot_text_file.write(snapshot_text)
-
-    # Throw an exception if the plan is not there.
-    if not os.path.exists(file_path):
-        raise FileNotFoundError(
-            f"Could not find snapshot file at path {file_path}. Re-run with --overwrite-snapshots and check git status "
-            "to see what's new."
-        )
-
-    if mf_test_configuration.display_snapshots:
-        if not mf_test_configuration.overwrite_snapshots:
-            logger.warning(f"Not overwriting snapshots, so displaying existing snapshot at {file_path}")
-
-        if len(request.session.items) > 1:
-            raise ValueError("Displaying snapshots is only supported when there's a single item in a testing session.")
-        webbrowser.open("file://" + file_path)
-
-    # Read the existing plan from the file and compare with the actual plan
-    with open(file_path, "r") as snapshot_text_file:
-        expected_snapshot_text = snapshot_text_file.read()
-
-        if exclude_line_regex:
-            # Filter out lines that should be ignored.
-            expected_snapshot_text = _exclude_lines_matching_regex(
-                file_contents=expected_snapshot_text, exclude_line_regex=exclude_line_regex
-            )
-            snapshot_text = _exclude_lines_matching_regex(
-                file_contents=snapshot_text, exclude_line_regex=exclude_line_regex
-            )
-        # pytest should show a detailed diff with "assert actual_modified == expected_modified", but it's not, so doing
-        # this instead.
-        if snapshot_text != expected_snapshot_text:
-            differ = difflib.Differ()
-            diff = differ.compare(expected_snapshot_text.splitlines(), snapshot_text.splitlines())
-            assert False, f"Snapshot from {file_path} does not match. Diff from expected to actual:\n" + "\n".join(diff)
 
 
 def assert_execution_plan_text_equal(  # noqa: D103
